app/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.6
sqlalchemy==2.0.38
psycopg[binary]==3.2.5
alembic==1.14.1
confluent-kafka==2.6.1
httpx==0.27.2
python-dotenv==1.0.1
prometheus-client==0.21.1
opentelemetry-api==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-exporter-otlp==1.30.0
opentelemetry-instrumentation-fastapi==0.51b0
opentelemetry-instrumentation-requests==0.51b0